대회 수정

+++ Assets/Ranking.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class Ranking
 {
     public int[] charm_type = new int[6];
-    public int[] charm_charmming = new int[6];
+    public double[] charm_charmming = new double[6];
     public int[] swim_type = new int[6];
-    public int[] swim_hp = new int[6];
+    public double[] swim_hp = new double[6];
     public int[] hurdle_type = new int[6];
-    public int[] hurdle_speedjump = new int[6];
+    public double[] hurdle_speedjump = new double[6];
 }
